* origin/pr/20683:
  Astyle
  Rewrite trap loading to use generic_factory

* origin/pr/20688:
  Remove files that have regressions in astyle 2.0.6

* origin/pr/20700:
  Fix starting NPC positioning
  Fix NPC positioning in mapgen

* origin/pr/20743:
  surface_heat in medsandchemicals
  water_boiling_heat in recipe_others
  surface_heat in recipe_others
  List-ify brewing.json

* origin/pr/20752:
  Update battery.json
  Update battery.json
  battery install time tweaks

* origin/pr/20760:
  Translation template update

* origin/pr/20762:
  Fix some MSVC compile warnings on implicit casting

* origin/pr/20763:
  Turn on 'ENCODING_CONV' by default

* origin/pr/20767:
  Added possibility to board locked doors

+++ src/character.cpp
@@ -2233,7 +2233,7 @@ std::string Character::extended_description() const
     const auto &bps = get_all_body_parts( true );
     // Find length of bp names, to align
     // accumulate looks weird here, any better function?
-    size_t longest = std::accumulate( bps.begin(), bps.end(), 0, []( size_t m, body_part bp ) {
+    size_t longest = std::accumulate( bps.begin(), bps.end(), ( size_t )0, []( size_t m, body_part bp ) {
         return std::max( m, body_part_name_as_heading( bp, 1 ).size() );
     } );
