+++ src/overmap.cpp
@@ -331,7 +331,7 @@ void overmap_specials::check_consistency()
     const size_t max_count = ( OMAPX / OMSPEC_FREQ ) * ( OMAPY / OMSPEC_FREQ ) / 2;
     const size_t actual_count = std::accumulate(  specials.get_all().begin(), specials.get_all().end(), 0,
     []( size_t sum, const overmap_special &elem ) {
-        return sum + ( elem.flags.count( "UNIQUE" ) == 0 ? std::max( elem.occurrences.min, 0 ) : 1 );
+        return sum + ( elem.flags.count( "UNIQUE" ) == ( size_t )0 ? ( size_t )std::max( elem.occurrences.min, 0 ) : ( size_t )1 );
     } );
 
     if( actual_count > max_count ) {
