+++ src/pickup.cpp
@@ -1057,7 +1057,7 @@ void Pickup::pick_up( const tripoint &pos, int min )
 
             if( it._item.count_by_charges() ) {
                 size_t num_picked = std::min( ( size_t )it._item.charges, count );
-                pick_values.push_back( { it.idx, num_picked } );
+                pick_values.push_back( { static_cast<int>( it.idx ), static_cast<int>( num_picked ) } );
                 count -= num_picked;
             } else {
                 size_t num_picked = 1;
